Readme revision
Most of the links in this file are also on the Contribute page of the website, so I trimmed them here and added a link. It's a pretty short read now.

Added descriptions of the files in this repo, as it may not be obvious what their purpose is.

Also updated a broken Slack invite link.