Fix date formatting on the event issue template
The template has been wrong, leading to events not being included in the calendar. Encountered this with three different people who created events in the past 1-2 weeks. Was wondering why I saw the issues but nothing pop up in my calendar tools. Turns out it was due to this template being wrong.