Calendar workflow fix
The workflow is currently run on the "main" branch, but should only run on the "calendar" branch. You can see the failed workflow runs in the "Actions" tab on the repo.

This PR limits the workflow runs to the "calendar" branch, and hopefully fixes the issue.

Not totally sure why this started happening only recently. My guess it that it's related to GitHub requiring an update to a newer Node version, which I made a few days ago in #437.